move namespace info to opc ua server stack
